Add table module for Solihull dashboard.
When tested against preview, this shows no data, because there is no data
for the most recent week in preview. So I've added some local data to test
against. The local data is taken from preview using the same query params
but with an earlier start date and end date.