Extend calculateLinearTicks to take a divisor argument
We will use this for displaying axis ticks in minutes, when
our raw data is in seconds.

If we use seconds to calculate ticks and then apply a minute
formatter afterwards, we end up with odd-looking ticks (e.g. at
1000 seconds, or 17 minutes). So we should allow
calculateLinearTicks to take an optional divisor argument. If it
is supplied, we get ticks formatted as minutes at nicely even
values.

Update tests.